requirements.txt
Flask
Flask-Cors
PyPDF2
SQLAlchemy
psycopg2-binary
google-generativeai